requirements.txt
torch>=2.0.1
diffusers>=0.21.0
transformers>=4.31.0
datasets>=2.14.0
accelerate>=0.21.0
peft>=0.6.0
torchvision>=0.15.0
pillow>=10.0.0
xformers>=0.0.20
git+https://github.com/richzhang/PerceptualSimilarity